Donation Wizard: Only allow 'one-off' donations for event fundraisers
- Previously we would always set the default donation to 'monthly',
for both charity donations and event fundraiser donations.

- Now, we continue to default to 'monthly' for charity donations, but
default to 'one-off' donations for event fundraisers, and stop showing
any repeat options for event fundraisers.

How it looks now:
- Event fundraiser donation wizard: https://imgur.com/a/OtkB4fM
- Charity donation wizard: https://imgur.com/a/OaSU2T7

- (Ideally in the Edit Fundraiser page you would be able to decide
whether to allow repeating donations, but this is a quick fix to
stop showing the monthly donate option for an existing fundraiser.)